Fixed the coding standards errors
