--- src/main/java/Utility/Ui.java
@@ -1,4 +1,0 @@
-package Utility;
-
-public class Ui {
-}
