+++ src/main/java/seedu/duke/utility/InputParser.java
@@ -1,4 +1,4 @@
-package Utility;
+package seedu.duke.utility;
 
 public class InputParser {
 }
