+++ src/main/java/seedu/duke/utility/SaveState.java
@@ -1,7 +1,8 @@
-package Utility;
+package seedu.duke.utility;
 
 
 public interface SaveState {
     void saveState();
+    
     void loadState();
 }
