+++ src/main/java/seedu/duke/utility/Ui.java
@@ -1,0 +1,4 @@
+package seedu.duke.utility;
+
+public class Ui {
+}
